Base code changes for module split
Going on the previous two commits bringing in exisitng code over to the
workspace structure, this makes it work together using each other as
dependencies and such + misc fixes so it compiles. This includes the
CIP25 crate as well.
