+++ .gitignore
@@ -6,8 +6,24 @@ rust/target
 rust/.idea
 binaryen/
 
-rust/core/target/**
-rust/core/Cargo.lock
-rust/wasm/target/**
-rust/wasm/Cargo.lock
-rust/json-gen-split/target/**
+target/
+
+core/target/**
+core/Cargo.lock
+
+crypto/target/**
+crypto/Cargo.lock
+
+wasm/target/**
+wasm/Cargo.lock
+
+cip25/target/**
+cip25/Cargo.lock
+
+cip25-wasm/target/**
+cip25-wasm/Cargo.lock
+
+json-gen-split/target/**
+
+cip25-json-gen/target/**
+
