+++ chain/Cargo.toml
@@ -1,0 +1,54 @@
+[package]
+name = "cardano-multiplatform-lib-chain"
+version = "0.1.0"
+edition = "2018"
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+cardano-multiplatform-lib-core = { "path" = "../core" }
+cardano-multiplatform-lib-crypto = { "path" = "../crypto" }
+cbor_event = "2.2.0"
+linked-hash-map = "0.5.3"
+derivative = "2.2.0"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0.57"
+schemars = "0.8.8"
+
+bech32 = "0.7.2"
+hex = "0.4.0"
+itertools = "0.10.1"
+getrandom = { version = "0.2.3", features = ["js"] }
+rand = "0.8.5"
+fraction = "0.10.0"
+base64 = "0.13"
+num-bigint = "0.4.0"
+num-integer = "0.1.45"
+#rand_os = "0.1"
+thiserror = "1.0.37"
+# These can be removed if we make wasm bindings for ALL functionality here.
+# This was not done right now as there is a lot of existing legacy code e.g.
+# for Byron that might need to be used from WASM and might not.
+# We can remove this dependency when that is decided.
+#
+# The other use-case here is enums. Without this two enums would need to be defined
+# despite wasm_bindgen supporting C-style enums (with non-negative values) 100%
+# This could possibly be resolved with macros but maybe not.
+
+# non-wasm
+#[target.'cfg(not(all(target_arch = "wasm32", not(target_os = "emscripten"))))'.dependencies]
+#rand_os = "0.1"
+#noop_proc_macro = "0.3.0"
+
+# wasm
+#[target.'cfg(all(target_arch = "wasm32", not(target_os = "emscripten")))'.dependencies]
+wasm-bindgen = { version = "=0.2.82", features = ["serde-serialize"] }
+#rand_os = { version = "0.1", features = ["wasm-bindgen"] }
+#js-sys = "=0.3.59"
+
+
+[dev-dependencies]
+quickcheck = "0.9.2"
+quickcheck_macros = "0.9.1"
+rand_chacha = "0.3.1"
