+++ chain/README.md
@@ -1,0 +1,4 @@
+# Chain
+
+This is the core cardano-multiplatform-lib crate for all on-chain data types.
+This was generated from the `specs/babbage/` CDDL specs using cddl-codegen.
