+++ chain/src/block.rs
@@ -1,3 +1,5 @@
+use super::*;
+
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema)]
 pub struct Block {
     pub header: Header,
