+++ chain/src/metadata.rs
@@ -1,3 +1,5 @@
+use super::*;
+
 #[derive(Clone, Debug, serde::Deserialize, serde::Serialize, schemars::JsonSchema)]
 pub struct AlonzoAuxData {
     pub key_0: Option<Metadata>,
