+++ cip25-wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cip25-lib-wasm"
+name = "cardano-multiplatform-lib-cip25-wasm"
 version = "0.1.0"
 edition = "2018"
 
@@ -7,7 +7,7 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-core = { path = "../core", package = "cip25-lib" }
+core = { path = "../cip25", package = "cardano-multiplatform-lib-cip25" }
 cbor_event = "2.2.0"
 wasm-bindgen = { version = "0.2", features=["serde-serialize"] }
 linked-hash-map = "0.5.3"
