+++ core/README.md
@@ -1,0 +1,3 @@
+# Core
+
+The most core types for cardano-multiplatform-lib. Used for core serialization types, errors, etc that will be used by other consuming crates within cardano-multiplatform-lib. End users should never have to use this crate directly.
