+++ crypto/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cardano-multiplatform-lib-cip25"
+name = "cardano-multiplatform-lib-crypto"
 version = "0.1.0"
 edition = "2018"
 
@@ -9,6 +9,17 @@ crate-type = ["cdylib", "rlib"]
 [dependencies]
 cardano-multiplatform-lib-core = { "path" = "../core" }
 cbor_event = "2.2.0"
+cryptoxide = "0.4.2"
+ed25519-bip32 = "0.4.1"
+sha2 = "^0.9"
+digest = "^0.9"
+bech32 = "0.7.2"
+hex = "0.4.0"
+thiserror = "1.0.37"
+rand = "0.8.5"
+cfg-if = "1"
+
+derivative = "2.2.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.57"
 schemars = "0.8.8"
