+++ crypto/README.md
@@ -1,0 +1,3 @@
+# Crypto
+
+Cryptography-related components for cardano-serialization-lib.
