+++ wasm/Cargo.toml
@@ -7,7 +7,8 @@ edition = "2018"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
-core = { path = "../core", package = "cardano-multiplatform-lib-core" }
+core = { path = "../chain", package = "cardano-multiplatform-lib-chain" }
+core_crypto = { path = "../crypto", package = "cardano-multiplatform-lib-crypto" }
 cbor_event = "2.2.0"
 wasm-bindgen = { version = "0.2", features=["serde-serialize"] }
 linked-hash-map = "0.5.3"
