+++ wasm/README.md
@@ -1,0 +1,3 @@
+# wasm
+
+WASM bindings covering the `chain` and `crypto` crates. We will need to try to split this off into separate bindings crates later if that is possible.
