+++ wasm/src/lib.rs
@@ -6,9 +6,11 @@ use crypto::*;
 
 use wasm_bindgen::prelude::{wasm_bindgen, JsValue, JsError};
 
+// this is actually chain pulling in the core definition.
+// when we regenerate we might want to change the module naming
 use core::{
     // this can now be directly exposed as it implements std::error::Error
-    error::DeserializeError,
+    DeserializeError,
 };
 
 #[wasm_bindgen]
